requirements.txt
requests 
pandas 
PyGithub 
groq